+++ app/src/main/java/com/tnco/runar/data/remote/BackgroundInfo.kt
@@ -1,6 +1,11 @@
-package com.tnco.runar.retrofit
+package com.tnco.runar.data.remote
 
 import android.graphics.Bitmap
 
-data class BackgroundInfo (val name: String, val imgUrl: String, var img: Bitmap?,var isSelected: Boolean = false){
+data class BackgroundInfo(
+    val name: String,
+    val imgUrl: String,
+    var img: Bitmap?,
+    var isSelected: Boolean = false
+) {
 }
