+++ app/src/main/java/com/tnco/runar/ui/activity/MainActivity.kt
@@ -1,9 +1,11 @@
 package com.tnco.runar.ui.activity
 
 import android.content.Context
+import android.content.Intent
 import android.content.IntentFilter
 import android.media.AudioManager
 import android.os.Bundle
+import android.view.View
 import android.view.WindowManager
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
