+++ app/src/main/java/com/tnco/runar/ui/component/dialog/CancelDialog.kt
@@ -13,7 +13,7 @@ import com.tnco.runar.R
 import com.tnco.runar.analytics.AnalyticsHelper
 import com.tnco.runar.enums.AnalyticsEvent
 import com.tnco.runar.ui.Navigator
-import com.tnco.runar.utils.AnalyticsConstants
+import com.tnco.runar.util.AnalyticsConstants
 
 class CancelDialog(
     private val context: Context,
