+++ app/src/main/java/com/tnco/runar/ui/fragment/FavouriteFragment.kt
@@ -33,8 +33,8 @@ import com.tnco.runar.enums.AnalyticsEvent
 import com.tnco.runar.ui.Navigator
 import com.tnco.runar.ui.component.dialog.SavedLayoutsDialog
 import com.tnco.runar.ui.viewmodel.FavouriteViewModel
-import com.tnco.runar.utils.AnalyticsConstants
-import com.tnco.runar.utils.AnalyticsUtils
+import com.tnco.runar.util.AnalyticsConstants
+import com.tnco.runar.util.AnalyticsUtils
 
 class FavouriteFragment : Fragment() {
     val viewModel: FavouriteViewModel by viewModels()
