+++ app/src/main/java/com/tnco/runar/ui/fragment/GeneratorMagicRune.kt
@@ -1,4 +1,4 @@
-package com.tnco.runar.ui.fragments
+package com.tnco.runar.ui.fragment
 
 import android.content.Intent
 import android.net.Uri
@@ -13,12 +13,12 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import com.tnco.runar.R
-import com.tnco.runar.controllers.AnalyticsHelper
-import com.tnco.runar.controllers.MusicController
+import com.tnco.runar.analytics.AnalyticsHelper
 import com.tnco.runar.enums.AnalyticsEvent
-import com.tnco.runar.presentation.viewmodel.MainViewModel
+import com.tnco.runar.feature.MusicController
 import com.tnco.runar.ui.activity.MainActivity
-import com.tnco.runar.utils.AnalyticsConstants
+import com.tnco.runar.ui.viewmodel.MainViewModel
+import com.tnco.runar.util.AnalyticsConstants
 import kotlinx.coroutines.delay
 
 class GeneratorMagicRune : Fragment() {
