+++ app/src/main/java/com/tnco/runar/ui/fragment/GeneratorStartFragment.kt
@@ -19,8 +19,8 @@ import com.tnco.runar.databinding.FragmentGeneratorStartBinding
 import com.tnco.runar.enums.AnalyticsEvent
 import com.tnco.runar.util.observeOnce
 import com.tnco.runar.model.RunesItemsModel
-import com.tnco.runar.presentation.viewmodel.MainViewModel
 import com.tnco.runar.ui.activity.MainActivity
+import com.tnco.runar.ui.viewmodel.MainViewModel
 import kotlinx.coroutines.launch
 import java.util.*
