+++ app/src/main/java/com/tnco/runar/ui/fragment/LayoutDescriptionFragment.kt
@@ -14,8 +14,8 @@ import com.tnco.runar.enums.AnalyticsEvent
 import com.tnco.runar.util.setOnCLickListenerForAll
 import com.tnco.runar.ui.viewmodel.DescriptionViewModel
 import com.tnco.runar.ui.Navigator
-import com.tnco.runar.utils.AnalyticsConstants
-import com.tnco.runar.utils.AnalyticsUtils
+import com.tnco.runar.util.AnalyticsConstants
+import com.tnco.runar.util.AnalyticsUtils
 
 class LayoutDescriptionFragment : Fragment(R.layout.fragment_layout_description),
     View.OnClickListener {
