+++ app/src/main/java/com/tnco/runar/ui/fragment/LayoutProcessingFragment.kt
@@ -17,8 +17,8 @@ import com.tnco.runar.databinding.FragmentLayoutProcessingBinding
 import com.tnco.runar.enums.AnalyticsEvent
 import com.tnco.runar.ui.viewmodel.ProcessingViewModel
 import com.tnco.runar.ui.Navigator
-import com.tnco.runar.utils.AnalyticsConstants
-import com.tnco.runar.utils.AnalyticsUtils
+import com.tnco.runar.util.AnalyticsConstants
+import com.tnco.runar.util.AnalyticsUtils
 import kotlinx.coroutines.delay
