+++ app/src/main/java/com/tnco/runar/ui/fragment/LibraryFragment.kt
@@ -31,7 +31,7 @@ import com.tnco.runar.R
 import com.tnco.runar.analytics.AnalyticsHelper
 import com.tnco.runar.enums.AnalyticsEvent
 import com.tnco.runar.ui.viewmodel.LibraryViewModel
-import com.tnco.runar.utils.AnalyticsConstants
+import com.tnco.runar.util.AnalyticsConstants
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
