+++ app/src/main/java/com/tnco/runar/ui/fragment/RunePatternGenerator.kt
@@ -1,4 +1,4 @@
-package com.tnco.runar.ui.fragments
+package com.tnco.runar.ui.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -9,10 +9,10 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.tnco.runar.R
-import com.tnco.runar.controllers.AnalyticsHelper
+import com.tnco.runar.analytics.AnalyticsHelper
 import com.tnco.runar.enums.AnalyticsEvent
-import com.tnco.runar.presentation.viewmodel.MainViewModel
 import com.tnco.runar.ui.activity.MainActivity
+import com.tnco.runar.ui.viewmodel.MainViewModel
 
 class RunePatternGenerator : Fragment() {
     private lateinit var viewModel: MainViewModel
