+++ app/src/main/java/com/tnco/runar/ui/viewmodel/FavouriteViewModel.kt
@@ -8,8 +8,8 @@ import com.tnco.runar.enums.AnalyticsEvent
 import com.tnco.runar.model.*
 import com.tnco.runar.repository.DatabaseRepository
 import com.tnco.runar.repository.SharedDataRepository
-import com.tnco.runar.utils.AnalyticsConstants
-import com.tnco.runar.utils.AnalyticsUtils
+import com.tnco.runar.util.AnalyticsConstants
+import com.tnco.runar.util.AnalyticsUtils
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
