+++ app/src/main/res/layout/fragment_generator_processing.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.tnco.runar.customView.OffsetImageView
+    <com.tnco.runar.ui.component.view.OffsetImageView
         android:id="@+id/generator_offsetImageView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
